Handle every case for strto[ld]
strtol and strtod can try to parse a number from a command line
argument that is an empty string, a string that has characters
other than what numbers have, it can overflow or underflow and
it can be too high or low for a specific use case.
Handle all these cases properly.